+++ changelog.md
@@ -4,6 +4,7 @@
 
 ## Changes in release 6.4.0 ##
 + Improved API key validation
++ Added support for negative payment fees
 
 ## Changes in release 6.3.0 ##
 + Apple Pay Direct fixed